users/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from users.views import UserViewSet

app_name = "user"

router = DefaultRouter()
router.register(r"", UserViewSet, basename="user")

urlpatterns = [path("", include(router.urls))]
