requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-storages
boto3
Pillow
drf-yasg
django-filter
django-cors-headers
gunicorn
whitenoise
dj-database-url
setuptools
resend
django-easy-audit
mypy
djangorestframework-stubs[compatible-mypy]
